+++ ntex/src/web/middleware/defaultheaders.rs
@@ -114,7 +114,6 @@ where
     type Res = WebResponse;
     type Error = S::Error;
 
-    crate::forward_poll!(service);
     crate::forward_ready!(service);
     crate::forward_shutdown!(service);